+++ bin/gen-doc.sh
@@ -3,9 +3,14 @@
 
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"  # this file dirname
+
+# types chart
 convey --disable-external --show-uml 0 | dot -Tsvg -o $DIR/../docs/convey-methods.svg
 HELPFILE=$DIR/../docs/convey-help-cmd-output.md
 echo "This is the output of the \`--help\` command." > $HELPFILE
 echo "\`\`\`" >> $HELPFILE
 convey --help >> $HELPFILE
 echo "\`\`\`" >> $HELPFILE
+
+# autocompletion file
+convey --get-autocompletion > $DIR/../extra/convey-autocompletion.bash