+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 * LICENSE included
 * daemon – since it takes around 0.75 s to start the program, notably because of the external libraries (pint, bs4, requests), when doing a single query or a headless processing (no menu involved) we may reduce this time at least ten times with a daemon)
 * emergency input mode – when piping into the program (instead of giving the input as an argument), convey tries to gain a reduced STDIN from the terminal process #38
+* bash completion
 * fix:
     - adding external modules at runtime via menu
     - Ctrl+C works when interrupting wizzard as expected #39
     - detect multiline quoted_printable
+* internal:
+    - daemon pipes stderr as well
 
 ## 1.2 (2019-11-13)
 * web service